+++ ZFSin/zfs/lib/libspl/include/sys/list_impl.h
@@ -39,8 +39,8 @@ struct list_node {
 };
 
 struct list {
-	size_t	list_size;
-	size_t	list_offset;
+	unsigned int	list_size;
+	unsigned int	list_offset;
 	struct list_node list_head;
 };
